fix: resolve N+1 query on stars table (#7200)
